+++ src/sections/MainPage/index.module.scss
@@ -1,21 +1,35 @@
 .main-page--profile-placement {
-  transform: translate(-50%, -50%);
-  top: 50vh;
-  left: 30vw;
+	top: 50vh;
+	left: 30vw;
 }
 
 .main-page--profile-picture {
-  height: 300px;
-  width: 300px;
+	height: 300px;
+	width: 300px;
 }
 
-.main-page--logo-placement {
-  height: 300px;
-  width: 300px;
-}
+.main-page--logo {
+  height: 70px;
 
-.gmail-logo {
-  fill: #ff0000;
-  top: 50vh;
-  left: 60vw;
+	&-placement {
+		&-github {
+			left: 55vw;
+			top: 50vh;
+		}
+
+		&-gmail {
+			left: 65vw;
+			top: 50vh;
+		}
+
+		&-linkedin {
+			left: 75vw;
+			top: 50vh;
+		}
+
+		&-medium {
+			left: 85vw;
+			top: 50vh;
+		}
+	}
 }
