+++ src/sections/MainPage/index.tsx
@@ -4,38 +4,58 @@ import classnames from "classnames";
 import styles from "./index.module.scss";
 
 import ProfilePicture from "../../assets/profile_picture.jpg";
-import { Gmail, GitHub, Medium, LinkedIn } from "../../components/SVGs/Icons";
+import { Gmail, GitHub, Medium, LinkedIn } from "../../components/Icons";
 
 export const MainPage: FunctionComponent = () => {
-  return (
-    <>
-      <div>
-        <img
-          src={ProfilePicture}
-          alt=""
-          className={classnames(
-            "ui-circle",
-            "ui-position--fixed",
-            styles["main-page--profile-placement"],
-            styles["main-page--profile-picture"]
-          )}
-        />
-      </div>
-      <div>
-        <div
-          className={classnames(
-            "ui-circle",
-            "ui-position--fixed",
-            "ui-overflow--hidden",
-            styles["main-page--logo-placement"]
-          )}
-        >
-          <Gmail />
-        </div>
-        <GitHub />
-        <Medium />
-        <LinkedIn />
-      </div>
-    </>
-  );
+	return (
+		<>
+			<div>
+				<img
+					src={ProfilePicture}
+					alt=""
+					className={classnames(
+						"ui-circle",
+						"ui-position--fixed",
+						"ui-axis--normalize",
+						styles["main-page--profile-placement"],
+						styles["main-page--profile-picture"]
+					)}
+				/>
+			</div>
+			<div>
+				<GitHub
+					className={classnames(
+						"ui-position--fixed",
+						"ui-axis--normalize",
+						styles["main-page--logo"],
+						styles["main-page--logo-placement-github"]
+					)}
+				/>
+				<Gmail
+					className={classnames(
+						"ui-position--fixed",
+						"ui-axis--normalize",
+						styles["main-page--logo"],
+						styles["main-page--logo-placement-gmail"]
+					)}
+				/>
+				<LinkedIn
+					className={classnames(
+						"ui-position--fixed",
+						"ui-axis--normalize",
+						styles["main-page--logo"],
+						styles["main-page--logo-placement-linkedin"]
+					)}
+				/>
+				<Medium
+					className={classnames(
+						"ui-position--fixed",
+						"ui-axis--normalize",
+						styles["main-page--logo"],
+						styles["main-page--logo-placement-medium"]
+					)}
+				/>
+			</div>
+		</>
+	);
 };
