+++ src/styles/colors.scss
@@ -5,67 +5,89 @@ $charcoal: /**/ #230f0f;
 $olive: /*   */ #3f3a3a;
 
 .ui-bg {
-  &--white {
-    background-color: $white;
-  }
+	&--white {
+		background-color: $white;
+	}
 
-  &--black {
-    background-color: $black;
-  }
+	&--black {
+		background-color: $black;
+	}
 
-  &--red {
-    background-color: $red;
-  }
+	&--red {
+		background-color: $red;
+	}
 
-  &--charcoal {
-    background-color: $charcoal;
-  }
+	&--charcoal {
+		background-color: $charcoal;
+	}
 
-  &--olive {
-    background-color: $olive;
-  }
+	&--olive {
+		background-color: $olive;
+	}
 }
 
 .ui-color {
-  &--white {
-    color: $white;
-  }
+	&--white {
+		color: $white;
+	}
 
-  &--black {
-    color: $black;
-  }
+	&--black {
+		color: $black;
+	}
 
-  &--red {
-    color: $red;
-  }
+	&--red {
+		color: $red;
+	}
 
-  &--charcoal {
-    color: $charcoal;
-  }
+	&--charcoal {
+		color: $charcoal;
+	}
 
-  &--olive {
-    color: $olive;
-  }
+	&--olive {
+		color: $olive;
+	}
 }
 
 .ui-border-color {
-  &--white {
-    border-color: $white;
-  }
+	&--white {
+		border-color: $white;
+	}
 
-  &--black {
-    border-color: $black;
-  }
+	&--black {
+		border-color: $black;
+	}
 
-  &--red {
-    border-color: $red;
-  }
+	&--red {
+		border-color: $red;
+	}
 
-  &--charcoal {
-    border-color: $charcoal;
-  }
+	&--charcoal {
+		border-color: $charcoal;
+	}
 
-  &--olive {
-    border-color: $olive;
-  }
+	&--olive {
+		border-color: $olive;
+	}
+}
+
+.ui-padding-color {
+	&--white {
+		border-color: $white;
+	}
+
+	&--black {
+		border-color: $black;
+	}
+
+	&--red {
+		border-color: $red;
+	}
+
+	&--charcoal {
+		border-color: $charcoal;
+	}
+
+	&--olive {
+		border-color: $olive;
+	}
 }
