+++ src/styles/positions.scss
@@ -37,3 +37,7 @@
     }
   }
 }
+
+.ui-axis--normalize {
+  transform: translate(-50%, -50%);
+}
